tools/requirements.txt
agate==1.12.0
agate-dbf==0.2.3
agate-excel==0.4.1
agate-sql==0.7.2
babel==2.16.0
certifi==2024.8.30
click==8.1.7
dbfread==2.0.7
et_xmlfile==2.0.0
greenlet==3.1.1
isodate==0.7.2
leather==0.4.0
numpy==2.1.3
olefile==0.47
openpyxl==3.1.5
parsedatetime==2.6
psycopg2-binary==2.9.10
pyproj==3.7.0
python-dotenv==1.0.1
python-slugify==8.0.4
pytimeparse==1.1.8
shapely==2.0.6
SQLAlchemy==2.0.36
text-unidecode==1.3
typing_extensions==4.12.2
xlrd==2.0.1
